refactor: split instrument sampler out of tile machinery
InstrumentSample and resolveSample move to instrument_sample.dart. They
are shared with the Perdix media overlay, which resolves a sample per
video frame, so they must survive the removal of the tile preference
machinery in the fullscreen profile redesign (#811).

Pure move: no behavior change.
