+++ lib/features/media/presentation/widgets/perdix_overlay/perdix_face_resolver.dart
@@ -3,7 +3,7 @@ import 'package:submersion/features/dive_log/data/services/profile_analysis_serv
 import 'package:submersion/features/dive_log/domain/entities/dive.dart';
 import 'package:submersion/features/dive_log/domain/entities/gas_switch.dart';
 import 'package:submersion/features/dive_log/domain/services/profile_position.dart';
-import 'package:submersion/features/dive_log/presentation/widgets/instrument_tiles.dart';
+import 'package:submersion/features/dive_log/presentation/widgets/instrument_sample.dart';
 
 /// Snapshot of dive computer readings at one moment of the dive, shaped for
 /// the Perdix-style media overlay. Values are metric; formatting happens in
