+++ test/features/dive_log/presentation/widgets/instrument_tiles_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:submersion/core/constants/enums.dart';
 import 'package:submersion/core/deco/ascent_rate_calculator.dart';
 import 'package:submersion/core/deco/entities/o2_exposure.dart';
 import 'package:submersion/features/dive_log/data/services/profile_analysis_service.dart';
@@ -18,16 +17,6 @@ Dive _recDive() => Dive(
       depth: 10,
       temperature: i.isEven ? 20 : null,
     ),
-  ),
-);
-
-/// Recreational dive with no temperature data at all.
-Dive _recDiveNoTemp() => Dive(
-  id: 'd3',
-  dateTime: DateTime(2026, 1, 1, 10),
-  profile: List.generate(
-    7,
-    (i) => DiveProfilePoint(timestamp: i * 10, depth: 10),
   ),
 );
 
@@ -228,96 +217,6 @@ void main() {
         InstrumentTileId.ceiling,
         InstrumentTileId.tts,
       ]);
-    });
-  });
-
-  group('resolveSample', () {
-    test('reads point fields and curve values at the derived index', () {
-      final sample = resolveSample(
-        profile: _techDive().profile,
-        analysis: _techAnalysis(),
-        tankPressures: _techTankPressures(),
-        timestamp: 300,
-      );
-      expect(sample.depthMeters, isNotNull);
-      expect(sample.runtimeSeconds, 300);
-      expect(sample.tankPressuresBar, isNotEmpty);
-      expect(sample.ppO2Bar, 1.2);
-      expect(sample.ceilingMeters, 3.0);
-      expect(sample.ttsSeconds, (60 - 30) * 10);
-    });
-
-    test('null-at-position values stay null (temperature gap)', () {
-      final sample = resolveSample(
-        profile: _recDiveNoTemp().profile,
-        analysis: null,
-        timestamp: 60,
-      );
-      expect(sample.temperatureCelsius, isNull);
-    });
-
-    test('inDeco reflects decoType at the position', () {
-      final sample = resolveSample(
-        profile: _techDive().profile,
-        analysis: _techAnalysis(),
-        timestamp: 300,
-      );
-      expect(sample.inDeco, isTrue);
-    });
-
-    test('not in deco before the deco boundary', () {
-      final sample = resolveSample(
-        profile: _techDive().profile,
-        analysis: _techAnalysis(),
-        timestamp: 0,
-      );
-      expect(sample.inDeco, isFalse);
-      expect(sample.ndlSeconds, 600);
-    });
-
-    test('empty profile returns a bare sample keyed on the timestamp', () {
-      final sample = resolveSample(
-        profile: const [],
-        analysis: null,
-        timestamp: 42,
-      );
-      expect(sample.runtimeSeconds, 42);
-      expect(sample.depthMeters, isNull);
-      expect(sample.tankPressuresBar, isEmpty);
-      expect(sample.inDeco, isFalse);
-    });
-
-    test('curve values align to the passed profile late in the dive '
-        '(the profile must be the analysis basis, not dive.profile)', () {
-      // A short, coarsely-sampled active-source profile: 5 points at 60s
-      // intervals. The analysis curves are index-aligned to THIS array.
-      final profile = List.generate(
-        5,
-        (i) => DiveProfilePoint(timestamp: i * 60, depth: 10),
-      );
-      final analysis = ProfileAnalysis.empty().copyWith(
-        cnsCurve: [1.0, 2.0, 3.0, 4.0, 5.0],
-        ascentRates: [
-          for (var i = 0; i < 5; i++)
-            AscentRatePoint(
-              timestamp: i * 60,
-              depth: 10,
-              rateMetersPerMin: i.toDouble(),
-              category: AscentRateCategory.safe,
-            ),
-        ],
-      );
-
-      // Late in the dive: index 4 of the profile. Before the fix the index
-      // came from dive.profile (a longer, differently-sampled array), which
-      // read wrong values here and null past the curves' end.
-      final sample = resolveSample(
-        profile: profile,
-        analysis: analysis,
-        timestamp: 240,
-      );
-      expect(sample.cnsPercent, 5.0);
-      expect(sample.ascentRateMetersPerMin, 4.0);
     });
   });
 }
